ref(get recipes api): add type definitions, refactor getRecipeService and helper functions
